refactor: move SearchState type in state.ts
